+++ test/libsolidity/syntaxTests/comments/multiline_unicode_direction_override_5.sol
@@ -1,0 +1,17 @@
+// SPDX-License-Identifier: UNLICENSED
+pragma solidity ^0.7.0;
+
+contract C {
+    function f() public pure
+    {
+        // first push 1, then pop 1
+        /*ok ‮‬*/
+
+        // first push 2, then pop 2
+        /*ok ‮‮‬‬*/
+
+        // first push 3, then pop 3
+        /*ok ‮‮‮‬‬‬*/
+    }
+}
+// ----