+++ test/libsolidity/syntaxTests/comments/singleline_unicode_direction_override_1.sol
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: UNLICENSED
+pragma solidity ^0.7.0;
+
+contract C {
+    function f() public pure
+    {
+        // pop 1
+        // underflow ‬
+    }
+}
+// ----
+// ParserError 8936: (137-153): Unicode direction override underflow in comment.